requirements.txt
streamlit
torch
torchvision
matplotlib
scikit-learn
seaborn
Pillow
